requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
gitpython>=3.1.40
PyGithub>=2.3.0
pytest>=8.0.0
pytest-timeout>=2.3.0
streamlit>=1.35.0
pandas>=2.0.0
pyflakes>=3.2.0
bcrypt>=4.0.0
langchain>=0.2.0
